Fractional sensativities are more regular.
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@781 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/client/in_win.c
@@ -927,7 +927,7 @@ static void ProcessMouse(mouse_t *mouse, usercmd_t *cmd, int pnum)
 	extern int mousemove_x, mousemove_y;
 
 	int mx, my;
-	int mouse_x, mouse_y;
+	double mouse_x, mouse_y;
 
 	int i;
